requirements.txt
numpy
tensorflow
keras
opencv-python
pillow
PyQt5
scipy